+++ crates/forge-runner/src/running/config_run.rs
@@ -15,6 +15,7 @@ use cheatnet::runtime_extensions::forge_config_extension::{
 use runtime::{ExtendedRuntime, StarknetRuntime, starknet::context::build_context};
 use starknet_api::block::{
     BlockInfo, BlockNumber, BlockTimestamp, GasPrice, GasPriceVector, GasPrices, NonzeroGasPrice,
+    StarknetVersion,
 };
 use starknet_types_core::felt::Felt;
 use std::default::Default;
@@ -83,6 +84,7 @@ pub fn run_config_pass(
         },
         sequencer_address: 0_u8.into(),
         use_kzg_da: true,
+        starknet_version: StarknetVersion::LATEST,
     };
     let mut context = build_context(&block_info, None, &TrackedResource::from(tracked_resource));